toolkit/components/passwordmgr/test/mochitest/test_bug_627616.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test bug 627616 related to proxy authentication</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<script class="testbody" type="text/javascript">
    SimpleTest.waitForExplicitFinish();

    var Ci = SpecialPowers.Ci;

    function makeXHR(expectedStatus, expectedText, extra) {
      var xhr =  new XMLHttpRequest();
      xhr.open("GET", "authenticate.sjs?" +
                      "proxy_user=proxy_user&" +
                      "proxy_pass=proxy_pass&" +
                      "proxy_realm=proxy_realm&" +
                      "user=user1name&" +
                      "pass=user1pass&" +
                      "realm=mochirealm&" +
                      extra || "");
      xhr.onloadend = function() {
        is(xhr.status, expectedStatus, "xhr.status");
        is(xhr.statusText, expectedText, "xhr.statusText");
        runNextTest();
      };
      return xhr;
    }

    function testNonAnonymousCredentials() {
      var xhr = makeXHR(200, "OK");
      xhr.send();
    }

    function testAnonymousCredentials() {
      // Test that an anonymous request correctly performs proxy authentication
      var xhr = makeXHR(401, "Authentication required");
      SpecialPowers.wrap(xhr).channel.loadFlags |= Ci.nsIChannel.LOAD_ANONYMOUS;
      xhr.send();
    }

    function testAnonymousNoAuth() {
      // Next, test that an anonymous request still does not include any non-proxy
      // authentication headers.
      var xhr = makeXHR(200, "Authorization header not found", "anonymous=1");
      SpecialPowers.wrap(xhr).channel.loadFlags |= Ci.nsIChannel.LOAD_ANONYMOUS;
      xhr.send();
    }

    var gExpectedDialogs = 0;
    var gCurrentTest;
    function runNextTest() {
      is(gExpectedDialogs, 0, "received expected number of auth dialogs");
      mm.sendAsyncMessage("prepareForNextTest");
      mm.addMessageListener("prepareForNextTestDone", function prepared(_msg) {
        mm.removeMessageListener("prepareForNextTestDone", prepared);
        if (pendingTests.length) {
          ({expectedDialogs: gExpectedDialogs,
            test: gCurrentTest} = pendingTests.shift());
          gCurrentTest.call(this);
        } else {
          mm.sendAsyncMessage("cleanup");
          mm.addMessageListener("cleanupDone", () => {
            // mm.destroy() is called as a cleanup function by runInParent(), no
            // need to do it here.
            SimpleTest.finish();
          });
        }
      });
    }

    var pendingTests = [{expectedDialogs: 2, test: testNonAnonymousCredentials},
                        {expectedDialogs: 1, test: testAnonymousCredentials},
                        {expectedDialogs: 0, test: testAnonymousNoAuth}];

    const mm = runInParent(() => {
      const { classes: parentCc, interfaces: parentCi } = Components;

      const {NetUtil} = ChromeUtils.importESModule(
        "resource://gre/modules/NetUtil.sys.mjs"
      );

      const channel = NetUtil.newChannel({
        uri: "http://example.com",
        loadUsingSystemPrincipal: true,
      });

      const pps = parentCc["@mozilla.org/network/protocol-proxy-service;1"].
                getService(parentCi.nsIProtocolProxyService);
      pps.asyncResolve(channel, 0, {
        async onProxyAvailable(req, uri, pi, _status) {
          const mozproxy = "moz-proxy://" + pi.host + ":" + pi.port;
          const login1 = parentCc["@mozilla.org/login-manager/loginInfo;1"].
                      createInstance(parentCi.nsILoginInfo);
          login1.init(mozproxy, null, "proxy_realm", "proxy_user", "proxy_pass",
                     "", "");

          const login2 = parentCc["@mozilla.org/login-manager/loginInfo;1"].
                       createInstance(parentCi.nsILoginInfo);
          login2.init("http://mochi.test:8888", null, "mochirealm", "user1name",
                      "user1pass", "", "");
          await Services.logins.addLogins([login1, login2]);

          sendAsyncMessage("setupDone");
        },
        QueryInterface: ChromeUtils.generateQI([parentCi.nsIProtocolProxyCallback]),
      });

      addMessageListener("prepareForNextTest", _message => {
        parentCc["@mozilla.org/network/http-auth-manager;1"].
          getService(parentCi.nsIHttpAuthManager).
          clearAll();
        sendAsyncMessage("prepareForNextTestDone");
      });

      const dialogObserverTopic = "common-dialog-loaded";

      function dialogObserver(subj, _topic, _data) {
        subj.Dialog.ui.prompt.document
        .getElementById("commonDialog")
        .acceptDialog();
        sendAsyncMessage("promptAccepted");
      }

      Services.obs.addObserver(dialogObserver, dialogObserverTopic);

      addMessageListener("cleanup", _message => {
        Services.obs.removeObserver(dialogObserver, dialogObserverTopic);
        sendAsyncMessage("cleanupDone");
      });
    });

    mm.addMessageListener("promptAccepted", _message =>  gExpectedDialogs--);
    mm.addMessageListener("setupDone", _message => runNextTest());
</script>
</body>
</html>
